Accordion Progress
Thinking of getting rid of the accordion idea. As cool as an idea as it is I'm not able to make significant progress on the idea, plus see potential complications in the near future with trying to implement new data. Thinking that a better idea is to either bring up a new menu or pop up as it'll be easier to maintain. Plus lots of other apps already do it as well.