Deleted Day.png
After a bit of review of Color Psychology, I have decided that Day.png's design is lacking, and Sunset.png's design is actually really good for bringing out the emotions we are looking for. Therefor, I have Deleted Day.png, and the design for a dynamic packground has been changed to Sunset being the background to show that notificcations will be sent, and Night being the background to show that notifications are inactive. This will also be a subliminal way to enforce the user's sleep schedule if needed!